CLN: Move info (#31876)
